requirements.txt
pandas>=2.0.0
fastparquet>=2024.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
requests>=2.28.0
rapidfuzz>=3.0.0
python-dotenv>=1.0.0
tqdm>=4.64.0
psutil>=5.9.0
